+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/iteractionapi/feign/DeliveryClient.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import ru.yandex.practicum.iteractionapi.dto.DeliveryDto;
 import ru.yandex.practicum.iteractionapi.dto.OrderDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @FeignClient(name = "delivery", path = "/api/v1/delivery")
@@ -25,5 +26,5 @@ public interface DeliveryClient {
     void deliveryFailed(@RequestBody UUID deliveryId);
 
     @PostMapping("/cost")
-    Double deliveryCost(@RequestBody @Valid OrderDto orderDto);
+    BigDecimal deliveryCost(@RequestBody @Valid OrderDto orderDto);
 }
