+++ commerce/order/src/main/resources/schema.sql
@@ -9,9 +9,9 @@ CREATE TABLE IF NOT EXISTS orders
     delivery_weight  DOUBLE PRECISION,
     delivery_volume  DOUBLE PRECISION,
     fragile          BOOLEAN,
-    total_price      DOUBLE PRECISION,
-    delivery_price   DOUBLE PRECISION,
-    product_price    DOUBLE PRECISION
+    total_price      NUMERIC(19,4),
+    delivery_price   NUMERIC(19,4),
+    product_price    NUMERIC(19,4)
 );
 
 CREATE TABLE IF NOT EXISTS order_items
